Restructured and abstracted AssetMap class as parent of collections
