+++ src/models/assets/Asset.py
