+++ src/models/collections/AssetMap.py
@@ -1,6 +1,6 @@
-from src.models.Asset import Asset
+from src.models.assets.Asset import Asset
 
-class Watchlist:
+class AssetMap:
     def __init__(self, asset: Asset):
         self.dictId = 1
         self.asset = {0: asset}
