+++ src/models/collections/SoldAssets.py
@@ -1,0 +1,5 @@
+# Assets historically owned by the user
+class SoldAssets(AssetMap):
+    def __init__(self, asset: Asset):
+        self.dictId = 1
+        self.asset = {0: asset}
